refactor: extract mxgraph code in a dedicated file
